+++ Kwtc.Persistence.Tests/Extensions/DataRecordExtensionsTests.cs
@@ -1,8 +1,7 @@
-namespace Kwtc.Persistence.Tests;
+namespace Kwtc.Persistence.Tests.Extensions;
 
 using System.Data;
-using FluentAssertions;
-using Moq;
+using Persistence.Extensions;
 
 public class DataRecordExtensionsTests
 {
