+++ Kwtc.Persistence.Tests/Usings.cs
@@ -1,1 +1,3 @@
+global using FluentAssertions;
+global using Moq;
 global using Xunit;
